Adding support to preserve annotations at stream and app level.
Currently, the query is simply appended to stream definitions. If the query has annotations then the query becomes invalid. Hence the fix. If someone wants to enable statistics for a query, s/he has to add an app level annotation @app:statistics(reporter = 'console') before the query. However, this is currently not supported as the query is directly appended to the stream definitions misplacing the annotation and renders the query invalid. The fix is to insert the stream definitions just before the query. Below is the sample siddhi query with annotations which can be used for testing the fix.

@app:statistics(reporter = 'console') 
@query(name = 'echo')
from inputStream 
 select * 
 insert into outputStream